services/geospatial-worker/requirements.txt
azure-storage-blob==12.19.1
azure-identity==1.16.1
psycopg2-binary==2.9.9
requests==2.32.3
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.0
